requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
fake-useragent==0.1.11
frozenlist==1.3.0
idna==3.3
lxml==4.8.0
multidict==6.0.2
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.8
yarl==1.7.2
